requirements.txt
flask
pypdf
google-generativeai
python-dotenv